DVDDemuxShoutcast

+++ xbmc/cores/VideoPlayer/DVDDemuxers/DVDDemuxShoutcast.h
@@ -48,7 +48,7 @@ public:
 
   bool SeekTime(int time, bool backwords = false, double* startpts = NULL);
   int GetStreamLength();
-  CDemuxStream* GetStream(int iStreamId);
+  CDemuxStream* GetStream(int64_t iStreamId);
   int GetNrOfStreams();
 
 protected:
